fix windows debugfast build and warning fix in SDCardUtil.cpp
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@3225 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/Common/Src/SDCardUtil.cpp
@@ -171,11 +171,11 @@ static int write_empty(FILE* file, u64 count)
 	count *= 512;
 	while (count > 0)
 	{
-		int len = sizeof(empty);
+		u64 len = sizeof(empty);
 		if (len > count)
 			len = count;
 
-		if ( fwrite(empty, 1, len, file) != (size_t)len )
+		if ( fwrite(empty, 1, (size_t)len, file) != (size_t)len )
 			return 1;
 
 		count -= len;
